requirements.txt
allure_python_commons==2.10.0
beautifulsoup4==4.11.1
openpyxl==3.0.10
pytest==7.1.3
requests==2.28.1
selenium==4.4.3
webdriver_manager==3.8.3
wget==3.2
